move to src/lib/loadConfigFrom.ts
